+++ Editor/Unity.Platforms.Build.DotsRuntime/IDotsRuntimeBuildModifier.cs
@@ -1,0 +1,9 @@
+using Newtonsoft.Json.Linq;
+
+namespace Unity.Platforms.Build.DotsRuntime
+{
+    public interface IDotsRuntimeBuildModifier : IBuildComponent
+    {
+        void Modify(JObject settingsJObject);
+    }
+}