+++ Editor/Unity.Platforms.Build.Internals/AssemblyInfo.cs
@@ -1,0 +1,5 @@
+﻿using System.Runtime.CompilerServices;
+[assembly:InternalsVisibleTo("Unity.Entities.Runtime.Build")]
+[assembly:InternalsVisibleTo("Unity.Platforms.Build.Common")]
+[assembly:InternalsVisibleTo("Unity.Platforms.Build.Classic")]
+[assembly:InternalsVisibleTo("Unity.Scenes.Editor")]