+++ Editor/Unity.Platforms.Build/AssemblyInfo.cs
@@ -1,0 +1,5 @@
+﻿using System.Runtime.CompilerServices;
+[assembly:InternalsVisibleTo("Unity.Platforms.Build.Common")]
+[assembly:InternalsVisibleTo("Unity.Platforms.Build.Editor")]
+[assembly:InternalsVisibleTo("Unity.Platforms.Build.Internals")]
+[assembly:InternalsVisibleTo("Unity.Platforms.Build.Tests")]