+++ Editor/Unity.Platforms.Build/IBuildSettingsComponent.cs
@@ -1,0 +1,11 @@
+using System;
+using System.ComponentModel;
+
+namespace Unity.Platforms.Build
+{
+    [EditorBrowsable(EditorBrowsableState.Never)]
+    [Obsolete("IBuildSettingsComponent has been renamed to IBuildComponent. (RemovedAfter 2020-04-13) (UnityUpgradable) -> IBuildComponent")]
+    public interface IBuildSettingsComponent
+    {
+    }
+}