+++ Editor/Unity.Platforms.Build/IRunInstance.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Unity.Platforms.Build
+{
+    public interface IRunInstance : IDisposable
+    {
+        bool IsRunning { get; }
+    }
+}