+++ Editor/Unity.Platforms/AssemblyInfo.cs
@@ -1,2 +1,8 @@
 ﻿using System.Runtime.CompilerServices;
-[assembly:InternalsVisibleTo("Unity.Platforms.EditorTests")]
+[assembly:InternalsVisibleTo("Unity.Platforms.Tests")]
+[assembly:InternalsVisibleTo("Unity.Platforms.Build")]
+[assembly:InternalsVisibleTo("Unity.Platforms.Build.Common")]
+[assembly:InternalsVisibleTo("Unity.Platforms.Build.Classic")]
+[assembly:InternalsVisibleTo("Unity.Platforms.Build.Editor")]
+[assembly:InternalsVisibleTo("Unity.Platforms.Build.DotsRuntime")]
+[assembly:InternalsVisibleTo("Unity.Platforms.Build.Tests")]